Cryptomem: store seed in OTP of the CPU
Each cryptomem needs a secret seed to initiate the secure communication.
This seed is stored in the cryptomem and additionally in the OTP of the
CPU (so both parties have a copy of the secret).

OTP is memory mapped, but needs a separate entry in the MPU configuration.
Currently only read-access is mapped - for initial write access add MPU
configuration or do with MPU disabled (in prodtest).