Handle locking up zones by 4 wrong PW entries
If the cryptomem password is entered wrong 4 times, the zone locks up for
good.

Symptom: already sending the password command fails.
If the command succeeds but results in PAC counting to 0, the zone is
marked as locked.

Don't invalidate zone_index always, it is ok to assume it does not change.

No need to wipe a zone that is locked - it will not work anyway.