for messages that require an uninitialized device check no PIN too
After wipe the device should have no PIN. But if a PIN is set,
the messages that would initialize the device cannot handle the PIN, so
make sure, there actually is none.

This is relevant for cryptomem. Storage based PIN does not care, you can
just overwrite the PIN if there was any.