Cryptomem: make sure all operations in storage can access the cryptomem
In order to get the AES key from cryptomem, the PIN needs to be send.
This is true even if no PIN is in use - in which case the default
PIN is needed.

Make sure the right PIN is provided and cryptomem is open.

Before cm_get_aes_key() would do this internally, but this cannot be
controlled and could lead to default PIN instead of right PIN
send accidentally.