cryptomem: if setting PIN with reset_device, do not deauthenticate
Problem: reset_device can set the PIN using storage_setPin().
Later, encrypt_and_store_mnemonic() needs to be able to access the
cryptomem. It would be highly unexpected for the customer to enter a PIN
right after he set it and it would break compatibility to existing
software.

So do not deauthenticate cryptomem after setting pin. Just set to uncached
so the software will request a PIN for the next command.