cm: make cm_get_remaining_PIN_attempts() more robust
If for some reason cm_get_zone_index() fails, reset crypto and retry.

This means that this function should never be used after successful
password exchange, as it deauthenticates in the error case.

But still better than wiping the device because we think the cryptomem
locked the zone while there are still password attempts left.