implement deferred PIN set in recovery and reset
If a PIN needs to be set, request it from the user at the beginning.
But only set it at the end of the operation. Otherwise if anything goes
wrong the device may be stuck in a state where it is not yet initialized
but PIN already set. Now the PIN is ignored if the operation is not
finished.

Without cryptomem this is not a big deal, but with cryptomem you would not
be able to continue without entering the PIN first - unexpected e.g. for
reset or recovery (and some more commands) and had to call "wipe device"
to unblock.